requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
SQLAlchemy==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
